Give the stats read its own module
buildManifest did two unrelated jobs. It read a stats file, deciding what
the builder actually emitted across webpack, rspack and Vite spellings,
and it rolled that up into the manifest the Index compares. The two fail
differently and are tested differently, but both were only observable
through the same function, so "rspack puts the real file name in
nameForCondition" and "a roll-up is path-sensitive" shared one 662-line
suite and every builder-shape test paid for a disk, a config directory
and a Storybook version to assert which module the graph rooted at.

readStatsGraph now answers the first question and buildManifest only
rolls up. The seam is the graph value: files, hashes, story files and the
unrecognized entries. Hashing stays inside the reader rather than being
lifted out for purity, because detectStoryFiles already depends on it —
telling a story file apart from a require-context glob is asking whether
there is a file on disk, which is what the hashes answer. Hiding that
ordering is most of what the module is for.

pruneSyntheticFiles stays private to manifest.ts and still runs last. A
synthetic node is a reachable member of the subtrees the roll-ups walk,
so it contributes a [path, ''] entry; pruning any earlier would move
every hash that walked through it. That was the one fact the suite could
not see, so it is pinned first: a synthetic leaf hanging off a story now
has to move that story's hash even though pruning erases it from the
serialized graph. Verified by moving the prune call up and watching it
fail.

countNodeModulesFiles moves too. Its own comment already said it is a
property of the builder's output rather than of the manifest, and the
noNodeModulesFiles bail is asking the stats, not the manifest.

No behaviour change: the serialized manifest is byte-identical over the
664-module fixture in bin-src/__mocks__, and the golden hashes are
untouched.

The builder-spelling tests move to statsGraph.test.ts and assert on the
graph — that the right edge exists and the right bytes were hashed —
rather than on a story hash two steps downstream. Where a pair of tests
proved an edge exists and then that its content reaches the roll-up, the
pair collapses: the second claim is graph.test.ts's job, and end to end
the leaf-inclusion and golden tests already cover it. The relocation and
order-independence tests stay end to end in manifest.hashing.test.ts,
because the claim there is about the whole pipeline being path-canonical,
which the primitives being path-sensitive does not imply.
