+++ node-src/lib/turbosnap/v2/manifest.hashing.test.ts
@@ -7,9 +7,12 @@ import {
   outOfGraph,
   projectRoot,
   resetDisk,
-  withAbsent,
 } from './__fixtures__/manifestFixtures';
-import { buildManifest, serializeManifest } from './manifest';
+import { buildManifest } from './manifest';
+
+// These suites are about how a hash responds: what moves it, what leaves it alone, and what stays
+// stable when the project moves. They run end-to-end through buildManifest on purpose — the claim is
+// about the whole pipeline, not about any one step. Builder spellings live in statsGraph.test.ts.
 
 beforeEach(resetDisk);
 
@@ -322,341 +325,5 @@ describe('buildManifest relocation stability', () => {
     const second = await buildManifest(backwards, projectRoot, outOfGraph);
 
     expect(second.storybookHash).toBe(first.storybookHash);
-  });
-});
-
-describe('buildManifest concatenated modules', () => {
-  // Webpack/rspack concatenate the story and its local imports into one module: the module name
-  // carries a ` + N modules` suffix and the real files live in `module.modules`.
-  const concatenatedStory: Stats = {
-    modules: [
-      {
-        id: 1,
-        name: '/repo/packages/ui/src/Button.stories.tsx + 1 modules',
-        modules: [
-          { name: '/repo/packages/ui/src/Button.stories.tsx' },
-          { name: '/repo/packages/ui/src/Button.tsx' },
-        ],
-        reasons: [{ moduleName: './storybook-stories.js' }],
-      },
-    ],
-  };
-
-  it('keys the story by its root file, stripping the concatenation suffix', async () => {
-    disk.current.fileHashes = {
-      '/repo/packages/ui/src/Button.stories.tsx': 'S',
-      '/repo/packages/ui/src/Button.tsx': 'B',
-    };
-    const manifest = await buildManifest(concatenatedStory, projectRoot, outOfGraph);
-
-    expect([...manifest.storyFileHashes.keys()]).toEqual(['./src/Button.stories.tsx']);
-  });
-
-  it('records each concatenated sub-file as a dependency of the root file', async () => {
-    disk.current.fileHashes = {
-      '/repo/packages/ui/src/Button.stories.tsx': 'S',
-      '/repo/packages/ui/src/Button.tsx': 'B',
-    };
-    const manifest = await buildManifest(concatenatedStory, projectRoot, outOfGraph);
-
-    expect([...(manifest.files.get('./src/Button.stories.tsx')?.dependencies ?? [])]).toContain(
-      './src/Button.tsx'
-    );
-  });
-
-  it('changes the story hash when a concatenated sub-file content changes', async () => {
-    disk.current.fileHashes = {
-      '/repo/packages/ui/src/Button.stories.tsx': 'S',
-      '/repo/packages/ui/src/Button.tsx': 'B1',
-    };
-    const before = await buildManifest(concatenatedStory, projectRoot, outOfGraph);
-
-    disk.current.fileHashes = {
-      '/repo/packages/ui/src/Button.stories.tsx': 'S',
-      '/repo/packages/ui/src/Button.tsx': 'B2',
-    };
-    const after = await buildManifest(concatenatedStory, projectRoot, outOfGraph);
-
-    expect(after.storyFileHashes.get('./src/Button.stories.tsx')).not.toBe(
-      before.storyFileHashes.get('./src/Button.stories.tsx')
-    );
-  });
-
-  // `storybook-builder-rsbuild` 3.3.0/3.3.1 fill `modules` with the record's require-contexts rather
-  // than its concatenated files, and omit the root file. Reading the root from `modules` then yields
-  // a glob, which has no file on disk, so the whole record is promoted to a story importer and every
-  // module that imports it becomes a story file.
-  const contextsInModules: Stats = {
-    modules: [
-      {
-        id: 1,
-        name: './node_modules/storybook/dist/csf/index.js + 11 modules',
-        modules: [{ name: './node_modules/storybook/dist/csf/*.js' }],
-        reasons: [{ moduleName: './storybook-config-entry.js' }],
-      },
-      {
-        id: 2,
-        name: './node_modules/storybook/dist/instrumenter/index.js',
-        reasons: [{ moduleName: './node_modules/storybook/dist/csf/index.js + 11 modules' }],
-      },
-    ],
-  };
-
-  it('roots a module at its own name when `modules` holds contexts rather than concatenated files', async () => {
-    // The glob has no file on disk, which is what would promote the record to a story importer if
-    // the root were read from `modules`.
-    await withAbsent(
-      (candidate) => candidate.includes('*.js'),
-      async () => {
-        disk.current.fileHashes = {
-          '/repo/packages/ui/node_modules/storybook/dist/csf/index.js': 'C',
-          '/repo/packages/ui/node_modules/storybook/dist/instrumenter/index.js': 'I',
-        };
-
-        const manifest = await buildManifest(contextsInModules, projectRoot, outOfGraph);
-
-        // The record keys by its own name, so it stays a real file rather than becoming a story
-        // importer, and the module it imports is not a story file.
-        expect(manifest.files.has('./node_modules/storybook/dist/csf/index.js')).toBe(true);
-        expect([...manifest.storyFileHashes.keys()]).toEqual([]);
-      }
-    );
-  });
-});
-
-describe('buildManifest unhashable paths', () => {
-  it('skips a module named after a directory rather than failing the build', async () => {
-    const story = '/repo/packages/ui/src/Button.stories.tsx';
-    // rspack names one record after a directory on `storybook-builder-rsbuild` 3.3.0/3.3.1. Reading
-    // it throws EISDIR, which would fail the whole manifest to an internalError bail.
-    const directory = '/repo/packages/ui/node_modules/@storybook/react/dist/';
-
-    disk.current.fileHashes = { [story]: 'S' };
-    const manifest = await buildManifest(
-      {
-        modules: [
-          { id: 1, name: story, reasons: [{ moduleName: './storybook-stories.js' }] },
-          { id: 2, name: story, modules: [{ name: directory }], reasons: [] },
-        ],
-      },
-      projectRoot,
-      outOfGraph
-    );
-
-    expect([...manifest.storyFileHashes.keys()]).toEqual(['./src/Button.stories.tsx']);
-    expect(manifest.files.has('./node_modules/@storybook/react/dist')).toBe(false);
-  });
-});
-
-describe('buildManifest suffix-equivalent story importer identity', () => {
-  it('serializes the same manifest when only the story-entry reason gains a concatenation suffix', async () => {
-    const story = '/repo/packages/ui/src/Button.stories.tsx';
-    const stats = (storyImporter: string): Stats => ({
-      modules: [{ id: 1, name: story, reasons: [{ moduleName: storyImporter }] }],
-    });
-    disk.current.fileHashes = { [story]: 'S' };
-
-    const plain = serializeManifest(
-      await buildManifest(stats('./storybook-stories.js'), projectRoot, outOfGraph)
-    );
-    const concatenated = serializeManifest(
-      await buildManifest(stats('./storybook-stories.js + 1 modules'), projectRoot, outOfGraph)
-    );
-
-    expect(concatenated).toEqual(plain);
-  });
-});
-
-describe('buildManifest concatenated modules with rspack-style child names', () => {
-  // rspack labels a concatenated child's `name` with the parent group name (e.g.
-  // `./Button.stories.tsx + 1 modules`) rather than the child's own file. The real path is only in
-  // `nameForCondition`. Without reading `nameForCondition` the child collapses onto the root file
-  // and its content is never hashed.
-  const rspackConcatenatedStory: Stats = {
-    modules: [
-      {
-        id: 1,
-        name: '/repo/packages/ui/src/Button.stories.tsx + 1 modules',
-        nameForCondition: '/repo/packages/ui/src/Button.stories.tsx',
-        modules: [
-          {
-            name: '/repo/packages/ui/src/Button.stories.tsx',
-            nameForCondition: '/repo/packages/ui/src/Button.stories.tsx',
-          },
-          {
-            // The buggy child: `name` is the group label, real file is in nameForCondition.
-            name: '/repo/packages/ui/src/Button.stories.tsx + 1 modules',
-            nameForCondition: '/repo/packages/ui/src/Button.tsx',
-          },
-        ],
-        reasons: [{ moduleName: './storybook-stories.js' }],
-      },
-    ],
-  };
-
-  it('recovers the concatenated child file from nameForCondition', async () => {
-    disk.current.fileHashes = {
-      '/repo/packages/ui/src/Button.stories.tsx': 'S',
-      '/repo/packages/ui/src/Button.tsx': 'B',
-    };
-    const manifest = await buildManifest(rspackConcatenatedStory, projectRoot, outOfGraph);
-
-    expect([...manifest.storyFileHashes.keys()]).toEqual(['./src/Button.stories.tsx']);
-    expect([...(manifest.files.get('./src/Button.stories.tsx')?.dependencies ?? [])]).toContain(
-      './src/Button.tsx'
-    );
-  });
-
-  it('changes the story hash when the concatenated child content changes', async () => {
-    disk.current.fileHashes = {
-      '/repo/packages/ui/src/Button.stories.tsx': 'S',
-      '/repo/packages/ui/src/Button.tsx': 'B1',
-    };
-    const before = await buildManifest(rspackConcatenatedStory, projectRoot, outOfGraph);
-
-    disk.current.fileHashes = {
-      '/repo/packages/ui/src/Button.stories.tsx': 'S',
-      '/repo/packages/ui/src/Button.tsx': 'B2',
-    };
-    const after = await buildManifest(rspackConcatenatedStory, projectRoot, outOfGraph);
-
-    expect(after.storyFileHashes.get('./src/Button.stories.tsx')).not.toBe(
-      before.storyFileHashes.get('./src/Button.stories.tsx')
-    );
-  });
-
-  it('publishes the same manifest when minimal stats re-emit concatenated files by usable name', async () => {
-    const story = '/repo/packages/ui/src/Button.stories.tsx';
-    const implementation = '/repo/packages/ui/src/Button.tsx';
-    const shimmedStats: Stats = {
-      modules: [
-        {
-          id: 1,
-          name: story,
-          reasons: [{ moduleName: './storybook-stories.js' }],
-        },
-        {
-          id: 2,
-          name: implementation,
-          reasons: [{ moduleName: story }],
-        },
-      ],
-    };
-    disk.current.fileHashes = { [story]: 'S', [implementation]: 'B' };
-
-    const full = serializeManifest(
-      await buildManifest(rspackConcatenatedStory, projectRoot, outOfGraph)
-    );
-    const shimmed = serializeManifest(await buildManifest(shimmedStats, projectRoot, outOfGraph));
-
-    expect(shimmed).toEqual(full);
-  });
-});
-
-describe('buildManifest missing names', () => {
-  it('skips reasons with a null moduleName without dropping the story', async () => {
-    const stats: Stats = {
-      modules: [
-        {
-          id: 1,
-          name: '/repo/packages/ui/src/Button.stories.tsx',
-          // An entry reason carries `moduleName: null`; the stories-entry reason must still apply.
-          reasons: [
-            { moduleName: null as unknown as string },
-            { moduleName: './storybook-stories.js' },
-          ],
-        },
-      ],
-    };
-    disk.current.fileHashes = { '/repo/packages/ui/src/Button.stories.tsx': 'S' };
-
-    const manifest = await buildManifest(stats, projectRoot, outOfGraph);
-
-    expect([...manifest.storyFileHashes.keys()]).toEqual(['./src/Button.stories.tsx']);
-  });
-
-  it('uses module.modules when module.name is absent', async () => {
-    const stats: Stats = {
-      modules: [
-        {
-          id: 1,
-          name: null as unknown as string,
-          modules: [
-            { name: '/repo/packages/ui/src/Button.stories.tsx' },
-            { name: '/repo/packages/ui/src/Button.tsx' },
-          ],
-          reasons: [{ moduleName: './storybook-stories.js' }],
-        },
-      ],
-    };
-    disk.current.fileHashes = {
-      '/repo/packages/ui/src/Button.stories.tsx': 'S',
-      '/repo/packages/ui/src/Button.tsx': 'B',
-    };
-
-    const manifest = await buildManifest(stats, projectRoot, outOfGraph);
-
-    expect([...manifest.storyFileHashes.keys()]).toEqual(['./src/Button.stories.tsx']);
-    expect([...(manifest.files.get('./src/Button.stories.tsx')?.dependencies ?? [])]).toContain(
-      './src/Button.tsx'
-    );
-  });
-});
-
-describe('buildManifest hashFiles skip branches', () => {
-  it('contributes the empty string to a story hash for a file missing on disk', async () => {
-    const story = '/repo/packages/ui/src/Button.stories.tsx';
-    const missing = '/repo/packages/ui/src/missing.ts';
-
-    // Every path but this one has a file on disk, which is what makes the skip branch the only
-    // difference between the two builds below.
-    await withAbsent(
-      (candidate) => candidate === missing,
-      async () => {
-        const stats: Stats = {
-          modules: [
-            { id: 1, name: story, reasons: [{ moduleName: './storybook-stories.js' }] },
-            { id: 2, name: missing, reasons: [{ moduleName: story }] },
-          ],
-        };
-
-        disk.current.fileHashes = { [story]: 'S', [missing]: 'WOULD-BE-A' };
-        const before = await buildManifest(stats, projectRoot, outOfGraph);
-
-        // Change the content hash the missing file *would* have if it were hashed. If the skip
-        // branch didn't treat it as contributing '', this toggle would change the story hash.
-        disk.current.fileHashes = { [story]: 'S', [missing]: 'WOULD-BE-B' };
-        const after = await buildManifest(stats, projectRoot, outOfGraph);
-
-        expect(after.storyFileHashes.get('./src/Button.stories.tsx')).toBe(
-          before.storyFileHashes.get('./src/Button.stories.tsx')
-        );
-      }
-    );
-  });
-});
-
-describe('buildManifest hashing failures', () => {
-  it('fails the build rather than returning a manifest missing a file it could not read', async () => {
-    // Unreadability is a bug, not an answer: a manifest built without those bytes would silently
-    // under-capture, so this propagates to the entry point and bails TurboSnap to v1.
-    const story = '/repo/packages/ui/src/Button.stories.tsx';
-    const unreadable = new Error(`Could not hash ${story}: EACCES: permission denied`);
-
-    let err: Error | undefined;
-    try {
-      await buildManifest(
-        { modules: [{ id: 1, name: story, reasons: [{ moduleName: './storybook-stories.js' }] }] },
-        projectRoot,
-        {
-          ...outOfGraph,
-          projectFiles: { ...outOfGraph.projectFiles, hashAll: () => Promise.reject(unreadable) },
-        }
-      );
-    } catch (error) {
-      err = error as Error;
-    }
-
-    expect(err?.message).toContain(story);
   });
 });
